# docs/swagger.yaml
basePath: /api/v1
definitions:
  entities.CountryCurreny:
    description: Object containing country and corresponding currency.
    properties:
      country:
        type: string
      currency:
        type: string
    required:
    - country
    - currency
    type: object
  entities.ExchangeGetParams:
    description: Parameters to retrieve an exchange rate for a transaction.
    properties:
      country_currency:
        $ref: '#/definitions/entities.CountryCurreny'
      id:
        type: string
    required:
    - country_currency
    - id
    type: object
  entities.ExchangeTransaction:
    description: Transaction with exchange rate information.
    properties:
      converted_value:
        type: number
      country_currency:
        $ref: '#/definitions/entities.CountryCurreny'
      efective_date:
        type: string
      rate:
        type: number
      transaction:
        $ref: '#/definitions/entities.Transaction'
    type: object
  entities.Transaction:
    description: Purchase transaction with value, date, and description.
    properties:
      created_at:
        type: string
      date:
        type: string
      description:
        maxLength: 50
        type: string
      id:
        type: string
      updated_at:
        type: string
      value:
        type: number
    required:
    - date
    - description
    - id
    - value
    type: object
  rest.ResponseError:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API to persist transactions and convert the value to currencies of
    other countries
  title: VR Exchange API
  version: "1.0"
paths:
  /exchange:
    post:
      consumes:
      - application/json
      description: Retrieves exchange transaction based on a given transaction ID
        and currency
      parameters:
      - description: Exchange request
        in: body
        name: exchange
        required: true
        schema:
          $ref: '#/definitions/entities.ExchangeGetParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ExchangeTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Get exchange value
      tags:
      - exchange
  /transaction:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of transactions
      parameters:
      - description: Number of items to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Transaction'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: List transactions
      tags:
      - transaction
    post:
      consumes:
      - application/json
      description: Saves a new transaction to the system
      parameters:
      - description: Transaction object
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/entities.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Create a new transaction
      tags:
      - transaction
  /transaction/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a transaction by its unique ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Delete transaction
      tags:
      - transaction
    get:
      consumes:
      - application/json
      description: Retrieves a transaction by its unique ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Get transaction by ID
      tags:
      - transaction
swagger: "2.0"